pyproject.toml
[tool.poetry]
name = "chmigrate"
version = "0.5.0"
description = "Utilites make ClickHouse migration"
authors = ["Vadim Statishin <statishin@gmail.com>"]
license = "BSD-4-Clause"
readme = "README.md"
keywords = [
    "ClickHouse",
    "ch",
    "migrate",
    "migration",
]
classifiers = [
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "clickhouse_migrate" },
]

[tool.poetry.scripts]
chmigrate = 'clickhouse_migrate.runner:run'


[tool.poetry.dependencies]
python = "^3.8"
asynch = ">=0.1.9"
python-dotenv = "^0.19.1"
Jinja2 = "^3.0.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"
black = "^21.10b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-version-plugin]
source = "git-tag"
